ref(credentials): Simplify command proxy provider tracking
Centralize command proxy provider matching so credential activation and auth recovery use the same rule. Track active plugin providers with a Set so resume state carries the uniqueness invariant directly.

Refs GH-319
